+++ src/BackInStock.php
@@ -4,8 +4,8 @@
  *
  * Back in stock Craft Commerce 2 plugin
  *
- * @link      https://www.mylesderham.dev/
- * @copyright Copyright (c) 2019 Myles Derham
+ * @link      https://www.mylesthe.dev/
+ * @copyright Copyright (c) 2019 Myles Beardsmore
  */
 
 namespace mediabeastnz\backinstock;
@@ -30,7 +30,7 @@ use yii\base\Event;
 /**
  * Class BackInStock
  *
- * @author    Myles Derham
+ * @author    Myles Beardsmore
  * @package   BackInStock
  *
  * @property  BackInStockServiceService $backInStockService